+++ api/core/file/tool_file_parser.py
@@ -1,12 +1,20 @@
+"""Compatibility bridge for legacy ``core.file.tool_file_parser`` imports."""
+
 from collections.abc import Callable
-from typing import TYPE_CHECKING
+from typing import Any
 
-if TYPE_CHECKING:
-    from core.tools.tool_file_manager import ToolFileManager
+from core.workflow.file import tool_file_parser as workflow_tool_file_parser
 
-_tool_file_manager_factory: Callable[[], "ToolFileManager"] | None = None
+_tool_file_manager_factory: Callable[[], Any] | None = None
 
 
-def set_tool_file_manager_factory(factory: Callable[[], "ToolFileManager"]):
+def set_tool_file_manager_factory(factory: Callable[[], Any]) -> None:
     global _tool_file_manager_factory
     _tool_file_manager_factory = factory
+    workflow_tool_file_parser.set_tool_file_manager_factory(factory)
+
+
+__all__ = [
+    "_tool_file_manager_factory",
+    "set_tool_file_manager_factory",
+]
