+++ api/core/workflow/node_events/node.py
@@ -3,10 +3,10 @@ from datetime import datetime
 
 from pydantic import Field
 
-from core.file import File
 from core.model_runtime.entities.llm_entities import LLMUsage
 from core.rag.entities.citation_metadata import RetrievalSourceMetadata
 from core.workflow.entities.pause_reason import PauseReason
+from core.workflow.file import File
 from core.workflow.node_events import NodeRunResult
 
 from .base import NodeEventBase
