+++ api/core/workflow/nodes/document_extractor/node.py
@@ -21,11 +21,11 @@ from docx.table import Table
 from docx.text.paragraph import Paragraph
 
 from configs import dify_config
-from core.file import File, FileTransferMethod, file_manager
 from core.helper import ssrf_proxy
 from core.variables import ArrayFileSegment
 from core.variables.segments import ArrayStringSegment, FileSegment
 from core.workflow.enums import NodeType, WorkflowNodeExecutionStatus
+from core.workflow.file import File, FileTransferMethod, file_manager
 from core.workflow.node_events import NodeRunResult
 from core.workflow.nodes.base.node import Node
