+++ api/core/workflow/nodes/knowledge_retrieval/knowledge_retrieval_node.py
@@ -30,7 +30,7 @@ from .exc import (
 )
 
 if TYPE_CHECKING:
-    from core.file.models import File
+    from core.workflow.file.models import File
     from core.workflow.runtime import GraphRuntimeState
 
 logger = logging.getLogger(__name__)
