+++ api/core/workflow/nodes/question_classifier/question_classifier_node.py
@@ -39,7 +39,7 @@ from .template_prompts import (
 )
 
 if TYPE_CHECKING:
-    from core.file.models import File
+    from core.workflow.file.models import File
     from core.workflow.runtime import GraphRuntimeState
