+++ api/core/workflow/nodes/trigger_webhook/node.py
@@ -2,12 +2,12 @@ import logging
 from collections.abc import Mapping
 from typing import Any
 
-from core.file import FileTransferMethod
 from core.variables.types import SegmentType
 from core.variables.variables import FileVariable
 from core.workflow.constants import SYSTEM_VARIABLE_NODE_ID
 from core.workflow.entities.workflow_node_execution import WorkflowNodeExecutionStatus
 from core.workflow.enums import NodeExecutionType, NodeType
+from core.workflow.file import FileTransferMethod
 from core.workflow.node_events import NodeRunResult
 from core.workflow.nodes.base.node import Node
 from factories import file_factory
