+++ api/core/workflow/utils/condition/processor.py
@@ -2,9 +2,9 @@ import json
 from collections.abc import Mapping, Sequence
 from typing import Literal, NamedTuple
 
-from core.file import FileAttribute, file_manager
 from core.variables import ArrayFileSegment
 from core.variables.segments import ArrayBooleanSegment, BooleanSegment
+from core.workflow.file import FileAttribute, file_manager
 from core.workflow.runtime import VariablePool
 
 from .entities import Condition, SubCondition, SupportedComparisonOperator
