+++ api/extensions/ext_storage.py
@@ -124,3 +124,6 @@ storage = Storage()
 
 def init_app(app: DifyApp):
     storage.init_app(app)
+    from core.app.workflow.file_runtime import bind_dify_workflow_file_runtime
+
+    bind_dify_workflow_file_runtime()
