+++ api/tests/unit_tests/core/workflow/nodes/llm/test_file_saver.py
@@ -6,10 +6,10 @@ import httpx
 import pytest
 from sqlalchemy import Engine
 
-from core.file import FileTransferMethod, FileType, models
 from core.helper import ssrf_proxy
 from core.tools import signature
 from core.tools.tool_file_manager import ToolFileManager
+from core.workflow.file import FileTransferMethod, FileType, models
 from core.workflow.nodes.llm.file_saver import (
     FileSaverImpl,
     _extract_content_type_and_extension,
