+++ api/tests/unit_tests/core/workflow/nodes/webhook/test_webhook_node.py
@@ -3,10 +3,10 @@ from unittest.mock import patch
 import pytest
 
 from core.app.entities.app_invoke_entities import InvokeFrom
-from core.file import File, FileTransferMethod, FileType
 from core.variables import FileVariable, StringVariable
 from core.workflow.entities.graph_init_params import GraphInitParams
 from core.workflow.entities.workflow_node_execution import WorkflowNodeExecutionStatus
+from core.workflow.file import File, FileTransferMethod, FileType
 from core.workflow.nodes.trigger_webhook.entities import (
     ContentType,
     Method,
