+++ api/tests/unit_tests/core/workflow/test_variable_pool.py
@@ -3,7 +3,6 @@ from collections import defaultdict
 
 import pytest
 
-from core.file import File, FileTransferMethod, FileType
 from core.variables import FileSegment, StringSegment
 from core.variables.segments import (
     ArrayAnySegment,
@@ -27,6 +26,7 @@ from core.variables.variables import (
     Variable,
 )
 from core.workflow.constants import CONVERSATION_VARIABLE_NODE_ID, ENVIRONMENT_VARIABLE_NODE_ID, SYSTEM_VARIABLE_NODE_ID
+from core.workflow.file import File, FileTransferMethod, FileType
 from core.workflow.runtime import VariablePool
 from core.workflow.system_variable import SystemVariable
 from factories.variable_factory import build_segment, segment_to_variable
